deleted test_foundry.py, modified test_foundry_gha.py, added local_only_tests.py
